@@ -4,9 +4,9 @@ declare(strict_types=1);
 namespace App\Entity;
 
 use Doctrine\Common\Collections\ArrayCollection;
-use Doctrine\Common\Collections\Collection;
 use Doctrine\ORM\Mapping as ORM;
 use Symfony\Bridge\Doctrine\Validator\Constraints\UniqueEntity;
+use Symfony\Component\Security\Core\User\PasswordAuthenticatedUserInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
 
 /**
@@ -15,19 +15,19 @@ use Symfony\Component\Security\Core\User\UserInterface;
  * @ORM\Entity(repositoryClass="App\Repository\RssUserRepository")
  * @UniqueEntity("email", message="Email already taken")
  */
-class RssUser implements UserInterface
+class RssUser implements PasswordAuthenticatedUserInterface, UserInterface
 {
     /**
      * @ORM\Id()
      * @ORM\GeneratedValue()
      * @ORM\Column(type="integer")
      */
-    private $id;
+    private int $id;
 
     /**
      * @ORM\Column(type="string", length=180, unique=true)
      */
-    private $email;
+    private string $email;
 
     /**
      * @ORM\Column(type="json")
@@ -35,7 +35,6 @@ class RssUser implements UserInterface
     private array $roles = [];
 
     /**
-     * @var string The hashed password
      * @ORM\Column(type="string")
      */
     private string $password;
@@ -43,12 +42,12 @@ class RssUser implements UserInterface
     /**
      * @ORM\OneToMany(targetEntity="App\Entity\Feed", mappedBy="rssUser", orphanRemoval=true)
      */
-    private Collection $feeds;
+    private iterable $feeds;
 
     /**
      * @ORM\Column(type="string", length=255)
      */
-    private $opmlToken;
+    private string $opmlToken;
 
     public function __construct()
     {
@@ -79,7 +78,7 @@ class RssUser implements UserInterface
      */
     public function getUsername(): string
     {
-        return (string) $this->email;
+        return $this->email;
     }
 
     /**
@@ -134,9 +133,9 @@ class RssUser implements UserInterface
     }
 
     /**
-     * @return Collection|Feed[]
+     * @return Feed[]
      */
-    public function getFeeds(): Collection
+    public function getFeeds(): iterable
     {
         return $this->feeds;
     }
