@@ -14,35 +14,25 @@ use Symfony\Component\Console\Style\SymfonyStyle;
 class UserCreateCommand extends Command
 {
     protected static $defaultName = 'user:create';
-    /**
-     * @var RssUserRepository
-     */
-    private RssUserRepository $userRepository;
-    /**
-     * @var PasswordStrengthValidator
-     */
-    private PasswordStrengthValidator $passwdValidator;
 
-    public function __construct(RssUserRepository $userRepository, PasswordStrengthValidator $passwdValidator)
-    {
-        $this->userRepository  = $userRepository;
-        $this->passwdValidator = $passwdValidator;
-
+    public function __construct(
+        private RssUserRepository $userRepository,
+        private PasswordStrengthValidator $passwdValidator
+    ) {
         parent::__construct();
     }
 
-    protected function configure()
+    protected function configure(): void
     {
-        $this
-            ->setDescription('Creates a RSS user');
+        $this->setDescription('Creates a RSS user');
     }
 
-    protected function execute(InputInterface $input, OutputInterface $output)
+    protected function execute(InputInterface $input, OutputInterface $output): string|int
     {
         $io = new SymfonyStyle($input, $output);
         $io->success('Create a RSS user');
 
-        $email = $io->ask('Please enter the user\'s email', null, static function ($answer) {
+        $email = $io->ask('Please enter the user\'s email', null, static function (string $answer) {
             if (filter_var($answer, FILTER_VALIDATE_EMAIL) === false) {
                 throw new RuntimeException('Not a valid email address');
             }
@@ -50,7 +40,7 @@ class UserCreateCommand extends Command
             return $answer;
         });
 
-        $password = $io->ask('Please enter the user\'s password', null, function ($answer) {
+        $password = $io->ask('Please enter the user\'s password', null, function (string $answer) {
             if ($this->passwdValidator->validate($answer) === false) {
                 throw new RuntimeException('Weak password, please re-enter');
             }
@@ -59,7 +49,7 @@ class UserCreateCommand extends Command
         });
 
         $makeAdmin = false;
-        if ($io->ask('Make this user admin? (y/n)', 'n', static function ($answer) {
+        if ($io->ask('Make this user admin? (y/n)', 'n', static function (string $answer) {
                 if ($answer !== 'y' && $answer !== 'n') {
                     throw new RuntimeException('Answer `y` or `n`');
                 }
