@@ -9,26 +9,26 @@ use Symfony\Component\Form\Extension\Core\Type\EmailType;
 use Symfony\Component\Form\Extension\Core\Type\PasswordType;
 use Symfony\Component\Form\FormBuilderInterface;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Validator\Constraints\Email;
+use Symfony\Component\Validator\Constraints\NotBlank;
 
 class RssUserType extends AbstractType
 {
-    public function buildForm(FormBuilderInterface $builder, array $options)
+    public function buildForm(FormBuilderInterface $builder, array $options): void
     {
         $builder
-            ->add('email', EmailType::class)
-//            ->add(
-//                'roles',
-//                ChoiceType::class, [
-//                    'choices' => [
-//                        'ROLE_ADMIN',
-//                        'ROLE_USER',
-//                    ],
-//                ]
-//            )
-            ->add('password', PasswordType::class);
+            ->add('email', EmailType::class, [
+                'constraints' => [
+                    new NotBlank(),
+                    new Email(),
+                ],
+            ])
+            ->add('password', PasswordType::class, [
+                'constraints' => [new NotBlank()],
+            ]);
     }
 
-    public function configureOptions(OptionsResolver $resolver)
+    public function configureOptions(OptionsResolver $resolver): void
     {
         $resolver->setDefaults([
             'data_class' => RssUser::class,
