@@ -5,10 +5,10 @@ namespace App\Repository;
 
 use App\Entity\RssUser;
 use Doctrine\Bundle\DoctrineBundle\Repository\ServiceEntityRepository;
-use Doctrine\Common\Persistence\ManagerRegistry;
+use Doctrine\Persistence\ManagerRegistry;
 use InvalidArgumentException;
 use Ramsey\Uuid\Uuid;
-use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
+use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 
 /**
  * @method RssUser|null find($id, $lockMode = null, $lockVersion = null)
@@ -18,16 +18,9 @@ use Symfony\Component\Security\Core\Encoder\UserPasswordEncoderInterface;
  */
 class RssUserRepository extends ServiceEntityRepository
 {
-    /**
-     * @var UserPasswordEncoderInterface
-     */
-    private UserPasswordEncoderInterface $passwordEncoder;
-
-    public function __construct(ManagerRegistry $registry, UserPasswordEncoderInterface $passwordEncoder)
+    public function __construct(ManagerRegistry $registry, private UserPasswordHasherInterface $passwordEncoder)
     {
         parent::__construct($registry, RssUser::class);
-
-        $this->passwordEncoder = $passwordEncoder;
     }
 
     public function makeUser(string $email, string $password, bool $makeAdmin): RssUser
@@ -50,7 +43,7 @@ class RssUserRepository extends ServiceEntityRepository
             ->setRoles($roles)
             ->setOpmlToken(Uuid::uuid4()->toString());
 
-        $user->setPassword($this->passwordEncoder->encodePassword($user, $password));
+        $user->setPassword($this->passwordEncoder->hashPassword($user, $password));
 
         $this->save($user);
 
